+++ R/crosstable.h.R
@@ -8,7 +8,8 @@ crosstableOptions <- if (requireNamespace('jmvcore')) R6::R6Class(
         initialize = function(
             vars = NULL,
             group = NULL,
-            sty = "nejm", ...) {
+            sty = "gtsummary",
+            excl = TRUE, ...) {
 
             super$initialize(
                 package='ClinicoPath',
@@ -31,29 +32,43 @@ crosstableOptions <- if (requireNamespace('jmvcore')) R6::R6Class(
                 "sty",
                 sty,
                 options=list(
+                    "arsenal",
+                    "finalfit",
+                    "gtsummary",
                     "nejm",
                     "lancet",
                     "hmisc"),
-                default="nejm")
+                default="gtsummary")
+            private$..excl <- jmvcore::OptionBool$new(
+                "excl",
+                excl,
+                default=TRUE)
 
             self$.addOption(private$..vars)
             self$.addOption(private$..group)
             self$.addOption(private$..sty)
+            self$.addOption(private$..excl)
         }),
     active = list(
         vars = function() private$..vars$value,
         group = function() private$..group$value,
-        sty = function() private$..sty$value),
+        sty = function() private$..sty$value,
+        excl = function() private$..excl$value),
     private = list(
         ..vars = NA,
         ..group = NA,
-        ..sty = NA)
+        ..sty = NA,
+        ..excl = NA)
 )
 
 crosstableResults <- if (requireNamespace('jmvcore')) R6::R6Class(
     inherit = jmvcore::Group,
     active = list(
-        text3 = function() private$.items[["text3"]]),
+        todo = function() private$.items[["todo"]],
+        tablestyle1 = function() private$.items[["tablestyle1"]],
+        tablestyle2 = function() private$.items[["tablestyle2"]],
+        tablestyle3 = function() private$.items[["tablestyle3"]],
+        tablestyle4 = function() private$.items[["tablestyle4"]]),
     private = list(),
     public=list(
         initialize=function(options) {
@@ -61,10 +76,45 @@ crosstableResults <- if (requireNamespace('jmvcore')) R6::R6Class(
                 options=options,
                 name="",
                 title="Cross Tables")
+            self$add(jmvcore::Html$new(
+                options=options,
+                name="todo",
+                title="To Do"))
+            self$add(jmvcore::Html$new(
+                options=options,
+                name="tablestyle1",
+                title="`Cross Table - ${group}`",
+                clearWith=list(
+                    "vars",
+                    "group"),
+                visible="(sty:arsenal)",
+                refs="arsenal"))
+            self$add(jmvcore::Html$new(
+                options=options,
+                name="tablestyle2",
+                title="`Cross Table - ${group}`",
+                clearWith=list(
+                    "vars",
+                    "group"),
+                visible="(sty:finalfit)",
+                refs="finalfit"))
+            self$add(jmvcore::Html$new(
+                options=options,
+                name="tablestyle3",
+                title="`Cross Table - ${group}`",
+                clearWith=list(
+                    "vars",
+                    "group"),
+                visible="(sty:gtsummary)",
+                refs="gtsummary"))
             self$add(jmvcore::Preformatted$new(
                 options=options,
-                name="text3",
-                title="Cross Table",
+                name="tablestyle4",
+                title="`Cross Table - ${group}`",
+                clearWith=list(
+                    "vars",
+                    "group"),
+                visible="(sty:nejm || sty:lancet || sty:hmisc)",
                 refs="tangram"))}))
 
 crosstableBase <- if (requireNamespace('jmvcore')) R6::R6Class(
@@ -96,12 +146,17 @@ crosstableBase <- if (requireNamespace('jmvcore')) R6::R6Class(
 #' # example will be added
 #'}
 #' @param data The data as a data frame.
-#' @param vars .
+#' @param vars variable in the rows
 #' @param group variable in the column
 #' @param sty .
+#' @param excl .
 #' @return A results object containing:
 #' \tabular{llllll}{
-#'   \code{results$text3} \tab \tab \tab \tab \tab a preformatted \cr
+#'   \code{results$todo} \tab \tab \tab \tab \tab a html \cr
+#'   \code{results$tablestyle1} \tab \tab \tab \tab \tab a html \cr
+#'   \code{results$tablestyle2} \tab \tab \tab \tab \tab a html \cr
+#'   \code{results$tablestyle3} \tab \tab \tab \tab \tab a html \cr
+#'   \code{results$tablestyle4} \tab \tab \tab \tab \tab a preformatted \cr
 #' }
 #'
 #' @export
@@ -109,7 +164,8 @@ crosstable <- function(
     data,
     vars,
     group,
-    sty = "nejm") {
+    sty = "gtsummary",
+    excl = TRUE) {
 
     if ( ! requireNamespace('jmvcore'))
         stop('crosstable requires jmvcore to be installed (restart may be required)')
@@ -127,7 +183,8 @@ crosstable <- function(
     options <- crosstableOptions$new(
         vars = vars,
         group = group,
-        sty = sty)
+        sty = sty,
+        excl = excl)
 
     analysis <- crosstableClass$new(
         options = options,